requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.24.8
botocore==1.27.8
celery==5.2.7
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.5
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.29
pyparsing==3.0.9
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
